Convert to Composite Action
Convert the original chipsalliance/chisel version of this reusable
workflow to a composite action.  (A reusable workflow is not capable of
being made into a reusable action.)

Remove specializations for Chisel and make this more generic, e.g., by
allowing for customization of the staging branch name.  Add documentation
in the action and in a README.md.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>